Refactor funcitons which fetch delegations
Extracted a code snippets which fetch the operators of:
* grantee,
* managed grantee,
* copied delegations, to seprate functions to be able to use these
functions in other places.
