+++ solidity/dashboard/src/services/tokens-page.service.js
@@ -1,15 +1,18 @@
 import web3Utils from "web3-utils"
-import { sub, gt } from "../utils/arithmetics.utils"
 import moment from "moment"
-import { tokenGrantsService } from "../services/token-grants.service"
 import {
   createManagedGrantContractInstance,
   CONTRACT_DEPLOY_BLOCK_NUMBER,
 } from "../contracts"
 import { ContractsLoaded, Web3Loaded } from "../contracts"
+import {
+  getOperatorsOfOwner,
+  getOperatorsOfCopiedDelegations,
+  getOperatorsOfGrantee,
+  getOperatorsOfManagedGrantee,
+} from "./token-staking.service"
+import { sub, gt } from "../utils/arithmetics.utils"
 import { isEmptyArray } from "../utils/array.utils"
-import { getOperatorsOfOwner } from "./token-staking.service"
-import { isSameEthAddress } from "../utils/general.utils"
 
 export const fetchTokensPageData = async () => {
   const web3 = await Web3Loaded
@@ -40,13 +43,18 @@ export const fetchTokensPageData = async () => {
     granteeDelegations,
     granteeUndelegations,
     granteeGrantsIds,
-  ] = await getGranteeDelegations(initializationPeriod, undelegationPeriod)
+  ] = await getGranteeDelegations(
+    yourAddress,
+    initializationPeriod,
+    undelegationPeriod
+  )
 
   const [
     managedGrantsDelegations,
     managedGrantsUndelegations,
     managedGrantsIds,
   ] = await getManagedGranteeDelegations(
+    yourAddress,
     initializationPeriod,
     undelegationPeriod
   )
@@ -215,110 +223,13 @@ const getOwnedDelegations = async (
   )
 }
 
-const getAllGranteeOperators = async (
-  granteeOperators,
-  grantIds,
-  isManagedGrant = false
-) => {
-  const {
-    stakingContract,
-    tokenStakingEscrow,
-    stakingPortBackerContract,
-  } = await ContractsLoaded
-
-  let escrowRedelegation = []
-  if (!isEmptyArray(grantIds)) {
-    escrowRedelegation = await tokenStakingEscrow.getPastEvents(
-      "DepositRedelegated",
-      {
-        fromBlock: CONTRACT_DEPLOY_BLOCK_NUMBER.tokenStakingEscrow,
-        filter: {
-          grantId: grantIds,
-        },
-      }
-    )
-  }
-
-  const newOperatorToGrantId = escrowRedelegation.reduce((reducer, event) => {
-    const {
-      returnValues: { newOperator, grantId },
-    } = event
-    reducer[newOperator] = grantId
-    return reducer
-  }, {})
-
-  const newOperators = escrowRedelegation.map((_) => _.returnValues.newOperator)
-  const obsoleteOperators = escrowRedelegation.map(
-    (_) => _.returnValues.previousOperator
-  )
-
-  let activeOperators = granteeOperators
-    .filter((operator) => !obsoleteOperators.includes(operator))
-    .concat(newOperators)
-
-  const operatorsOfPortBacker = isEmptyArray(activeOperators)
-    ? []
-    : await getOperatorsOfOwner(
-        stakingPortBackerContract.options.address,
-        activeOperators
-      )
-
-  // We want to skip copied delegations
-  activeOperators = activeOperators.filter(
-    (operator) => !operatorsOfPortBacker.includes(operator)
-  )
-
-  let operatorsDetailsMap = {}
-  if (!isEmptyArray(activeOperators)) {
-    operatorsDetailsMap = (operatorsDetailsMap = await stakingContract.getPastEvents(
-      "OperatorStaked",
-      {
-        fromBlock: CONTRACT_DEPLOY_BLOCK_NUMBER.stakingContract,
-        filter: { operator: activeOperators },
-      }
-    )).reduce(toOperator, {})
-  }
-
-  for (const operator of Object.keys(operatorsDetailsMap)) {
-    const grantId = newOperatorToGrantId.hasOwnProperty(operator)
-      ? newOperatorToGrantId[operator]
-      : null
-    operatorsDetailsMap[operator] = {
-      ...operatorsDetailsMap[operator],
-      isFromGrant: true,
-      isManagedGrant,
-      grantId,
-    }
-  }
-
-  return operatorsDetailsMap
-}
-
 const getManagedGranteeDelegations = async (
+  address,
   initializationPeriod,
   undelegationPeriod
 ) => {
-  const { grantContract } = await ContractsLoaded
-
-  const managedGrants = await tokenGrantsService.fetchManagedGrants()
-
-  const grantIds = managedGrants.map(({ grantId }) => grantId)
-
-  const operators = new Set()
-
-  for (const managedGrant of managedGrants) {
-    const { managedGrantContractInstance } = managedGrant
-    const granteeAddress = managedGrantContractInstance.options.address
-    const grenteeOperators = await grantContract.methods
-      .getGranteeOperators(granteeAddress)
-      .call()
-    grenteeOperators.forEach(operators.add, operators)
-  }
-
-  const allOperators = await getAllGranteeOperators(
-    Array.from(operators),
-    grantIds,
-    true
+  const { allOperators, granteeGrants } = await getOperatorsOfManagedGrantee(
+    address
   )
 
   const [delegations, undelegations] = await getDelegations(
@@ -327,42 +238,15 @@ const getManagedGranteeDelegations = async (
     undelegationPeriod
   )
 
-  return [delegations, undelegations, grantIds]
+  return [delegations, undelegations, granteeGrants]
 }
 
 const getGranteeDelegations = async (
+  address,
   initializationPeriod,
   undelegationPeriod
 ) => {
-  const web3 = await Web3Loaded
-  const yourAddress = web3.eth.defaultAccount
-  const { grantContract } = await ContractsLoaded
-
-  // `getGrants` function returns grants for a grant manager or grantee.
-  // So it's possible that the provided address is a grantee in grant A and a grant manager in grant B.
-  // In that case a `getGrants` function returns [A, B].
-  const grantIds = new Set(
-    await grantContract.methods.getGrants(yourAddress).call()
-  )
-
-  // Filter out grants. We just want grants from the grantee's perspective
-  const granteeGrants = []
-  for (const grantId of grantIds) {
-    const { grantee } = await grantContract.methods.getGrant(grantId).call()
-    if (isSameEthAddress(grantee, yourAddress)) {
-      granteeGrants.push(grantId)
-    }
-  }
-
-  const granteeOperators = new Set(
-    await grantContract.methods.getGranteeOperators(yourAddress).call()
-  )
-
-  const allOperators = await getAllGranteeOperators(
-    Array.from(granteeOperators),
-    granteeGrants,
-    false
-  )
+  const { allOperators, granteeGrants } = await getOperatorsOfGrantee(address)
 
   const [delegations, undelegations] = await getDelegations(
     allOperators,
@@ -379,36 +263,22 @@ export const getCopiedDelegations = async (
   initializationPeriod,
   undelegationPeriod
 ) => {
-  const {
-    stakingPortBackerContract,
-    grantContract,
-    stakingContract,
-  } = await ContractsLoaded
+  const { grantContract, stakingContract } = await ContractsLoaded
 
-  const operatorsToCheck = (
-    await stakingPortBackerContract.getPastEvents("StakeCopied", {
-      fromBlock: CONTRACT_DEPLOY_BLOCK_NUMBER.stakingPortBackerContract,
-      filter: { owner: ownerOrGrantee },
-    })
-  ).map((_) => _.returnValues.operator)
+  const operators = await getOperatorsOfCopiedDelegations(ownerOrGrantee)
 
-  // No delegations
-  if (isEmptyArray(operatorsToCheck)) {
+  // No delegations.
+  if (isEmptyArray(operators)) {
     return await getDelegations({}, initializationPeriod, undelegationPeriod)
   }
 
-  // We only want operators for whom the delegation has not been paid back.
-  const operatorsOfPortBacker = await getOperatorsOfOwner(
-    stakingPortBackerContract.options.address,
-    operatorsToCheck
-  )
-
-  // Scan `OperatorStaked` event by operator(indexed param) to get authorizer and beneficiary.
-  const operatorToDetails = (isEmptyArray(operatorsOfPortBacker)
+  // Scan `OperatorStaked` event by operator(indexed param) to get authorizer
+  // and beneficiary.
+  const operatorToDetails = (isEmptyArray(operators)
     ? []
     : await stakingContract.getPastEvents("OperatorStaked", {
         fromBlock: CONTRACT_DEPLOY_BLOCK_NUMBER.stakingContract,
-        filter: { operator: operatorsOfPortBacker },
+        filter: { operator: operators },
       })
   ).reduce((reducer, _) => {
     reducer[_.returnValues.operator] = _.returnValues
